fix scripts not being able to call other scripts
This issue is caused by the get_script function trying to acquire a lock which is already acquired when the get_next_action function of ScriptExecutionContext tries to call get_script. This appears to be a new issue on focal/python3, but from what I can see there was no change to threading.Lock.

It is be possible to fix this by using threading.RLock instead but that may break other things.