fix configs having to be loaded twice before anything was actually loaded
fixes #21

This issue was caused by an extraneous else in the load config which would skip actually loading the requested config if there were any commands already loaded. Just removing the else is not enough because one of the preconditions later down the line is that there are no commands loaded. Because the remove_command function is only a scheduled removal, I added a short wait in after sending those commands. This wait may not be reliable if the network is slow, but any alternative solutions like checking if commands are scheduled for removal rather than just checking if they have already been removed is more difficult.