orders/info subscribers no longer use infinite queue size
I think this could cause issues with the deputy updates and maybe with the stale
orders issue. After disconnecting from the roscore wifi, deputy treeview shows
red on the deputies as expected. Previously, reconnecting seemed to take a long
time. With this I think the info messages come through quicker